+++ Assets/Scripts/Enemy/Attack/RangedAttack.cs
@@ -2,11 +2,11 @@ using UnityEngine;
 
 public class RangedAttack : EnemyAttackBase
 {
-    public override void AttackPlayer(Vector2 direct,GameObject projectile, int damage)
+    public override void AttackPlayer(Vector2 direct,GameObject projectile, float damage)
     {
         Instantiate(projectile, direct, Quaternion.identity);
     }
-    public override void AttackProtectedTarget(Vector2 direct, GameObject projectile, int damage)
+    public override void AttackProtectedTarget(Vector2 direct, GameObject projectile, float damage)
     {
         Instantiate(projectile, direct, Quaternion.identity);
     }
